removed 'height' & 'width' from the fabric JSON entries; now we should use custom_properties['ROWS'] & custom_properties['COLS']
